refactor(api): extract fetchDoli into its own module
